+++ src/main/java/com/onegini/examples/resourcegateway/config/DeviceApiConfig.java
@@ -3,6 +3,11 @@ package com.onegini.examples.resourcegateway.config;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
 @Component
 @ConfigurationProperties(prefix = "device.api")
 public class DeviceApiConfig {
@@ -11,28 +16,4 @@ public class DeviceApiConfig {
   private String username;
   private String password;
 
-
-  public String getServerRoot() {
-    return serverRoot;
-  }
-
-  public void setServerRoot(final String serverRoot) {
-    this.serverRoot = serverRoot;
-  }
-
-  public String getUsername() {
-    return username;
-  }
-
-  public void setUsername(final String username) {
-    this.username = username;
-  }
-
-  public String getPassword() {
-    return password;
-  }
-
-  public void setPassword(final String password) {
-    this.password = password;
-  }
 }