+++ src/main/java/com/onegini/examples/resourcegateway/config/TokenServerConfig.java
@@ -3,6 +3,11 @@ package com.onegini.examples.resourcegateway.config;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
 @Component
 @ConfigurationProperties(prefix = "resource.gateway.token-server")
 public class TokenServerConfig {
@@ -11,27 +16,4 @@ public class TokenServerConfig {
   private String clientSecret;
   private String baseUri;
 
-  public String getClientId() {
-    return clientId;
-  }
-
-  public void setClientId(final String clientId) {
-    this.clientId = clientId;
-  }
-
-  public String getClientSecret() {
-    return clientSecret;
-  }
-
-  public void setClientSecret(final String clientSecret) {
-    this.clientSecret = clientSecret;
-  }
-
-  public String getBaseUri() {
-    return baseUri;
-  }
-
-  public void setBaseUri(final String baseUri) {
-    this.baseUri = baseUri;
-  }
 }