+++ src/main/java/com/onegini/examples/resourcegateway/model/DecoratedUser.java
@@ -1,0 +1,10 @@
+package com.onegini.examples.resourcegateway.model;
+
+import lombok.Data;
+
+@Data
+public class DecoratedUser {
+
+  private String decoratedUserId;
+
+}