+++ src/main/java/com/onegini/examples/resourcegateway/model/DeviceList.java
@@ -1,0 +1,16 @@
+package com.onegini.examples.resourcegateway.model;
+
+import java.util.List;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class DeviceList {
+
+  private List<Device> devices;
+
+}