+++ src/main/java/com/onegini/examples/resourcegateway/model/TokenType.java
@@ -3,7 +3,6 @@ package com.onegini.examples.resourcegateway.model;
 public enum TokenType {
   DEFAULT,
   FINGER_PRINT,
-  FIDO,
   CUSTOM_AUTHENTICATOR,
   IMPLICIT_AUTHENTICATION
 }