+++ src/main/java/com/onegini/examples/resourcegateway/service/TokenTypeValidationService.java
@@ -2,8 +2,10 @@ package com.onegini.examples.resourcegateway.service;
 
 import static com.onegini.examples.resourcegateway.model.TokenType.IMPLICIT_AUTHENTICATION;
 
-import org.apache.commons.lang3.ArrayUtils;
+import java.util.Collection;
+
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 
 import com.onegini.examples.resourcegateway.model.TokenType;
 import com.onegini.examples.resourcegateway.model.exception.InvalidAccessTokenException;
@@ -11,23 +13,23 @@ import com.onegini.examples.resourcegateway.model.exception.InvalidAccessTokenEx
 @Service
 public class TokenTypeValidationService {
 
-  public void validateImplicitAuthenticationToken(final TokenType[] tokenTypes) {
+  public void validateImplicitAuthenticationToken(final Collection<TokenType> tokenTypes) {
     if (isNoImplicitAuthenticationToken(tokenTypes)) {
       throw new InvalidAccessTokenException("Token is not an implicit authentication access token");
     }
   }
 
-  public void validateNoImplicitAuthenticationToken(final TokenType[] tokenTypes) {
+  public void validateNoImplicitAuthenticationToken(final Collection<TokenType> tokenTypes) {
     if (isImplicitAuthenticationToken(tokenTypes)) {
       throw new InvalidAccessTokenException("Token is an implicit authentication access token");
     }
   }
 
-  private boolean isImplicitAuthenticationToken(final TokenType[] tokenTypes) {
-    return ArrayUtils.contains(tokenTypes, IMPLICIT_AUTHENTICATION);
+  private boolean isImplicitAuthenticationToken(final Collection<TokenType> tokenTypes) {
+    return CollectionUtils.containsInstance(tokenTypes, IMPLICIT_AUTHENTICATION);
   }
 
-  private boolean isNoImplicitAuthenticationToken(final TokenType[] tokenTypes) {
+  private boolean isNoImplicitAuthenticationToken(final Collection<TokenType> tokenTypes) {
     return !isImplicitAuthenticationToken(tokenTypes);
   }
 }