+++ src/main/java/com/onegini/examples/resourcegateway/web/ResourcesController.java
@@ -1,80 +1,98 @@
 package com.onegini.examples.resourcegateway.web;
 
+import static com.onegini.examples.resourcegateway.service.ScopeValidationService.SCOPE_APPLICATION_DETAILS;
+import static com.onegini.examples.resourcegateway.service.ScopeValidationService.SCOPE_READ;
+import static com.onegini.examples.resourcegateway.service.ScopeValidationService.SCOPE_WRITE;
 import static org.springframework.http.HttpHeaders.AUTHORIZATION;
 import static org.springframework.http.HttpStatus.OK;
 
-import javax.annotation.Resource;
-
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.onegini.examples.resourcegateway.model.ApplicationDetails;
-import com.onegini.examples.resourcegateway.model.DecoratedUserId;
+import com.onegini.examples.resourcegateway.model.DecoratedUser;
+import com.onegini.examples.resourcegateway.model.DeviceList;
+import com.onegini.examples.resourcegateway.model.FormDataWithFiles;
+import com.onegini.examples.resourcegateway.model.MultipartResponse;
 import com.onegini.examples.resourcegateway.model.TokenIntrospectionResult;
 import com.onegini.examples.resourcegateway.service.AccessTokenExtractor;
 import com.onegini.examples.resourcegateway.service.DeviceApiRequestService;
+import com.onegini.examples.resourcegateway.service.MultipartService;
 import com.onegini.examples.resourcegateway.service.ScopeValidationService;
 import com.onegini.examples.resourcegateway.service.TokenTypeValidationService;
 import com.onegini.examples.resourcegateway.service.tokenintrospection.TokenIntrospectionService;
 import com.onegini.examples.resourcegateway.util.DecoratedUserIdBuilder;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 
+@Slf4j
 @RestController
 @RequestMapping(value = "/resources")
+@RequiredArgsConstructor
 public class ResourcesController {
 
-  @Resource
-  private TokenIntrospectionService tokenIntrospectionService;
-  @Resource
-  private ScopeValidationService scopeValidationService;
-  @Resource
-  private DeviceApiRequestService deviceApiRequestService;
-  @Resource
-  private AccessTokenExtractor accessTokenExtractor;
-  @Resource
-  private TokenTypeValidationService tokenTypeValidationService;
+  private final TokenIntrospectionService tokenIntrospectionService;
+  private final ScopeValidationService scopeValidationService;
+  private final DeviceApiRequestService deviceApiRequestService;
+  private final MultipartService multipartService;
+  private final AccessTokenExtractor accessTokenExtractor;
+  private final TokenTypeValidationService tokenTypeValidationService;
 
-  @RequestMapping(value = "/devices", method = RequestMethod.GET)
-  public ResponseEntity<?> getDevices(@RequestHeader(AUTHORIZATION) final String authorizationHeader) {
-
-    final String accessToken = accessTokenExtractor.extractFromHeader(authorizationHeader);
-    final TokenIntrospectionResult tokenIntrospectionResult = tokenIntrospectionService.introspectAccessToken(accessToken);
-
-    scopeValidationService.validateReadScopeGranted(tokenIntrospectionResult.getScope());
-    tokenTypeValidationService.validateNoImplicitAuthenticationToken(tokenIntrospectionResult.getAmr());
+  @GetMapping(value = "/devices")
+  public ResponseEntity<DeviceList> getDevices(@RequestHeader(name = AUTHORIZATION, required = false) final String authorizationHeader) {
+    final TokenIntrospectionResult tokenIntrospectionResult = getTokenIntrospectionResultFromHeader(authorizationHeader);
+    validateScopeAndTokenType(tokenIntrospectionResult, SCOPE_READ);
 
     return deviceApiRequestService.getDevices(tokenIntrospectionResult.getSub());
   }
 
-  @RequestMapping(value = "/application-details", method = RequestMethod.GET)
-  public ResponseEntity<?> getApplicationDetails(@RequestHeader(AUTHORIZATION) final String authorizationHeader) {
-
-    final String accessToken = accessTokenExtractor.extractFromHeader(authorizationHeader);
-    final TokenIntrospectionResult tokenIntrospectionResult = tokenIntrospectionService.introspectAccessToken(accessToken);
-
-    scopeValidationService.validateApplicationDetailsScopeGranted(tokenIntrospectionResult.getScope());
-    tokenTypeValidationService.validateNoImplicitAuthenticationToken(tokenIntrospectionResult.getAmr());
+  @GetMapping(value = "/application-details")
+  public ResponseEntity<ApplicationDetails> getApplicationDetails(@RequestHeader(name = AUTHORIZATION, required = false) final String authorizationHeader) {
+    final TokenIntrospectionResult tokenIntrospectionResult = getTokenIntrospectionResultFromHeader(authorizationHeader);
+    validateScopeAndTokenType(tokenIntrospectionResult, SCOPE_APPLICATION_DETAILS);
 
     final ApplicationDetails applicationDetails = new ApplicationDetails(tokenIntrospectionResult.getAppIdentifier(), tokenIntrospectionResult.getAppPlatform(),
         tokenIntrospectionResult.getAppVersion());
 
-    return new ResponseEntity<Object>(applicationDetails, OK);
+    return new ResponseEntity<>(applicationDetails, OK);
   }
 
-  @RequestMapping(value = "/user-id-decorated", method = RequestMethod.GET)
-  public ResponseEntity<?> getDecoratedUserId(@RequestHeader(AUTHORIZATION) final String authorizationHeader) {
-
-    final String accessToken = accessTokenExtractor.extractFromHeader(authorizationHeader);
-    final TokenIntrospectionResult tokenIntrospectionResult = tokenIntrospectionService.introspectAccessToken(accessToken);
+  @GetMapping(value = "/user-id-decorated")
+  public ResponseEntity<DecoratedUser> getDecoratedUserId(@RequestHeader(name = AUTHORIZATION, required = false) final String authorizationHeader) {
+    final TokenIntrospectionResult tokenIntrospectionResult = getTokenIntrospectionResultFromHeader(authorizationHeader);
 
     tokenTypeValidationService.validateImplicitAuthenticationToken(tokenIntrospectionResult.getAmr());
 
-    final DecoratedUserId decoratedUserId = new DecoratedUserIdBuilder()
+    final DecoratedUser decoratedUser = new DecoratedUserIdBuilder()
         .withUserId(tokenIntrospectionResult.getSub())
         .build();
 
-    return new ResponseEntity<>(decoratedUserId, OK);
+    return new ResponseEntity<>(decoratedUser, OK);
+  }
+
+  @PostMapping(value = "/file-upload")
+  public ResponseEntity<MultipartResponse> fileUpload(@RequestHeader(name = AUTHORIZATION, required = false) final String authorizationHeader,
+                                                      @ModelAttribute final FormDataWithFiles formDataWithFiles) {
+    final TokenIntrospectionResult tokenIntrospectionResult = getTokenIntrospectionResultFromHeader(authorizationHeader);
+    validateScopeAndTokenType(tokenIntrospectionResult, SCOPE_WRITE);
+
+    final MultipartResponse multipartResponse = multipartService.parse(formDataWithFiles);
+
+    return new ResponseEntity<>(multipartResponse, OK);
+  }
+
+  private void validateScopeAndTokenType(final TokenIntrospectionResult tokenIntrospectionResult, final String requiredScope) {
+    scopeValidationService.validateScopeGranted(tokenIntrospectionResult.getScope(), requiredScope);
+    tokenTypeValidationService.validateNoImplicitAuthenticationToken(tokenIntrospectionResult.getAmr());
+  }
+
+  private TokenIntrospectionResult getTokenIntrospectionResultFromHeader(final String authorizationHeader) {
+    final String accessToken = accessTokenExtractor.extractFromHeader(authorizationHeader);
+    return tokenIntrospectionService.introspectAccessToken(accessToken);
   }
 }