+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
+## 3.1.0
+* A new endpoint is exposed for testing an HTTP form post with attachments
+* Upgraded to the V4 version of the Token Server device API.
+* Updated dependencies
+
 ## 3.0.0
-  * Updated the dependencies
-  * Use the new Onegini release process
-  * Upgraded to the V3 version of the Token Server device API.
+* Updated the dependencies
+* Use the new Onegini release process
+* Upgraded to the V3 version of the Token Server device API.